requirements.txt
fastapi[standard]
pyairtable
python-decouple